+++ CHANGELOG.md
@@ -18,6 +18,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
   `DT_NEEDED` shared objects of the given shared object, as described by
   [dlopen(3)] and at the exception of the cache file `/etc/ld.so.cache`
   (maintained by [ldconfig(8)]) that is unchecked yet
+- Use both ELF header and [dynamic loader][ld.so(8)] (`SONAME` and `ABI`) to
+  detect the path to the library to preload
 
 ### Removed